backend/ml/analytics-service/src/analytics_service.py
import grpc
from concurrent import futures
import tensorflow as tf
import numpy as np
import mlflow
from typing import Dict, Any, List

# Import generated protobuf code
from healthcare.analytics.v1 import analytics_pb2
from healthcare.analytics.v1 import analytics_pb2_grpc

class AnalyticsService(analytics_pb2_grpc.AnalyticsServiceServicer):
    def __init__(self):
        self.models = {}
        self.initialize_models()

    def initialize_models(self):
        """Initialize all ML models"""
        # Add models as needed
        pass

    def GetPatientRiskPredictions(self, request, context):
        """Get patient risk predictions"""
        try:
            # Implement prediction logic
            return analytics_pb2.GetPatientRiskPredictionsResponse()
        except Exception as e:
            context.set_code(grpc.StatusCode.INTERNAL)
            context.set_details(str(e))
            return analytics_pb2.GetPatientRiskPredictionsResponse()

    def GetResourceUtilizationPredictions(self, request, context):
        """Get resource utilization predictions"""
        try:
            # Implement prediction logic
            return analytics_pb2.GetResourceUtilizationPredictionsResponse()
        except Exception as e:
            context.set_code(grpc.StatusCode.INTERNAL)
            context.set_details(str(e))
            return analytics_pb2.GetResourceUtilizationPredictionsResponse()

    def GetAppointmentNoShowPredictions(self, request, context):
        """Get appointment no-show predictions"""
        try:
            # Implement prediction logic
            return analytics_pb2.GetAppointmentNoShowPredictionsResponse()
        except Exception as e:
            context.set_code(grpc.StatusCode.INTERNAL)
            context.set_details(str(e))
            return analytics_pb2.GetAppointmentNoShowPredictionsResponse()

def serve():
    server = grpc.server(futures.ThreadPoolExecutor(max_workers=10))
    analytics_pb2_grpc.add_AnalyticsServiceServicer_to_server(
        AnalyticsService(), server
    )
    server.add_insecure_port('[::]:50052')
    server.start()
    server.wait_for_termination()

if __name__ == '__main__':
    serve()
