backend/ml/analytics-service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
tensorflow==2.15.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3
mlflow==2.8.1
dvc==3.30.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
influxdb-client==1.36.1
